Fix: problems to get a Tag, Added new structure to defined standard values to modelManage
- Fixed error to get Tag on Panel component
- Initial implementation to define standard value to manageModelProvider
- Fixed some more little problems

+++ reach-goals-app/api/tag/service.js
@@ -186,6 +186,8 @@ const unlinkTagOnAssignment = async (tagID, assignmentID) => {
 }
 
 const unlinkAllTagOnAssignment = async (assignmentID) => {
+    if (!assignmentID) return
+
     try {
         return await prisma.tagOnAssignment.deleteMany({
             where: {
